SquirRL-Auditor/requirements.txt
# SquirRL-Auditor Dependencies

# Core ML/RL
stable-baselines3>=2.0.0
gymnasium>=0.28.0
torch>=2.0.0
numpy>=1.24.0

# Scientific Computing
scipy>=1.10.0
pymdptoolbox>=4.0b3

# Web Application
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Export
kaleido>=0.2.1
fpdf2>=2.7.0

# Development (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
